Update system tests for validate own chain to proper directories
