+++ test/functional/system/loader/validate_own_chain/first_round.js
@@ -15,10 +15,10 @@
 'use strict';
 
 const Promise = require('bluebird');
-const exceptions = require('../../../../helpers/exceptions');
-const blockVersion = require('../../../../logic/block_version.js');
-const queriesHelper = require('../../common/sql/queriesHelper.js');
-const localCommon = require('../common');
+const exceptions = require('../../../../../helpers/exceptions');
+const blockVersion = require('../../../../../logic/block_version.js');
+const queriesHelper = require('../../../common/sql/queriesHelper.js');
+const localCommon = require('../../common');
 
 describe('validateOwnChain', () => {
 	let library;
@@ -26,7 +26,7 @@ describe('validateOwnChain', () => {
 	let addTransactionsAndForgePromise;
 
 	localCommon.beforeBlock(
-		'lisk_functional_validate_own_chain_full_two_rounds',
+		'lisk_functional_validate_own_chain_first_round',
 		lib => {
 			library = lib;
 			Queries = new queriesHelper(lib, lib.db);
@@ -37,20 +37,20 @@ describe('validateOwnChain', () => {
 		}
 	);
 
-	describe('forge 3 rounds (303 blocks) with version = 0', () => {
+	describe('forge 101 blocks with version = 0', () => {
 		before(() => {
 			// Set current block version to 0
 			blockVersion.currentBlockVersion = 0;
 
-			// Not consider the genesis block
-			return Promise.mapSeries([...Array(101 * 3 - 1)], () => {
+			// 9 blocks with 1 genesis block
+			return Promise.mapSeries([...Array(100)], () => {
 				return addTransactionsAndForgePromise(library, [], 0);
 			});
 		});
 
-		it('blockchain should be at height 303', () => {
+		it('blockchain should be at height 101', () => {
 			const lastBlock = library.modules.blocks.lastBlock.get();
-			return expect(lastBlock.height).to.eql(303);
+			return expect(lastBlock.height).to.eql(101);
 		});
 
 		it('all blocks should have version = 0', () => {
@@ -63,7 +63,7 @@ describe('validateOwnChain', () => {
 			});
 		});
 
-		describe('increase block version = 1 and exceptions for height = 101', () => {
+		describe('increase block version = 1 and exceptions for height = 5', () => {
 			before(done => {
 				const __private = library.rewiredModules.loader.__get__('__private');
 
@@ -72,28 +72,15 @@ describe('validateOwnChain', () => {
 
 				// Set proper exceptions for blocks versions
 				exceptions.blockVersions = {
-					0: { start: 0, end: 101 },
+					0: { start: 0, end: 5 },
 				};
 
-				library.rewiredModules.loader.prototype.onBlockchainReady = () => {
+				library.modules.loader.onBlockchainReady = () => {
 					__private.loaded = true;
 					done();
 				};
 
 				__private.loadBlockChain();
-			});
-
-			it('blockchain should be at height 101', () => {
-				const lastBlock = library.modules.blocks.lastBlock.get();
-				return expect(lastBlock.height).to.eql(101);
-			});
-
-			it('remaining blocks have version = 0', () => {
-				return Queries.getAllBlocks().then(rows => {
-					_.each(rows, row => {
-						expect(row.version).to.be.equal(0);
-					});
-				});
 			});
 
 			describe('forge 5 more blocks', () => {
@@ -103,15 +90,25 @@ describe('validateOwnChain', () => {
 					});
 				});
 
-				it('blockchain should be at height 106', () => {
+				it('blockchain should be at height 10', () => {
 					const lastBlock = library.modules.blocks.lastBlock.get();
-					return expect(lastBlock.height).to.eql(106);
+					return expect(lastBlock.height).to.eql(10);
+				});
+
+				it('first 5 blocks should have version = 0', () => {
+					return Queries.getAllBlocks().then(rows => {
+						_.each(rows, row => {
+							if (row.height <= 5) {
+								expect(row.version).to.be.equal(0);
+							}
+						});
+					});
 				});
 
 				it('last 5 blocks should have version = 1', () => {
 					return Queries.getAllBlocks().then(rows => {
 						_.each(rows, row => {
-							if (row.height > 101) {
+							if (row.height > 5) {
 								expect(row.version).to.be.equal(1);
 							}
 						});
